# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CAT Backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/outfits": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "全ての衣装情報を取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outfits"
                ],
                "summary": "全衣装取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.OutfitResponse"
                            }
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "IDとパスワードで認証し、JWTトークンを返す。このJWTトークンは Authorization: Bearer \u003ctoken\u003e としてヘッダーに付加してください。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザーログイン",
                "parameters": [
                    {
                        "description": "ユーザーログインリクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.UserLoginResponse"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "新規ユーザーを登録し、IDとパスワードを返す",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザー登録",
                "parameters": [
                    {
                        "description": "ユーザー登録リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.UserRegisterResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.OutfitResponse": {
            "type": "object",
            "properties": {
                "asset_path": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "series_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "controllers.UserLoginRequest": {
            "type": "object",
            "required": [
                "id",
                "password"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "controllers.UserLoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "controllers.UserRegisterRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 4
                }
            }
        },
        "controllers.UserRegisterResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JTW": {
            "description": "Bearer: に続いてトークンを入力してください",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}